Created a materials package to store Materials manager, the default material and the default selected material.  Each material keeps a pointer to this package so that they can share Material_program objects internally which enable them to share vertex/fragment programs. Now that we have a package we can shift the creation of the default material, default selected material to the material.c. The gfx_create_material function can also go into material.c. Changed to using Option_table_add_set_Material_entry in many option tables.
Fixed apply_transformation for data.  It used to get the FE_region from the
node, which in the case of data actually returns the parent FE_region as this
is the FE_region upon which the FE_fields are defined.  To avoid this I now
pass the FE_region down to the iterator routine.


git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@2841 3c1deb5b-d424-0410-962d-aba41a686d42