Added a new coordinate system, NORMALISED_WINDOW_COORDINATES.  This is a rectangular cartesian system and so does not modify the coordinates, but when displayed the coordinates are used relative so window space rather than model coordinate space.  This is implemented by addding a coordinate system to graphics_objects and using this to temporarily disable the matrix transforms for these graphics.
Added a per-compilation custom rendering function to glyphs.  This function is
currently passed a bounding block of coordinates for the label_field.  The
first use of this function and values is to create labeled ticks that make a graph axis.

Added code for iso_surface decimation and a decimation_threshold parameter to
the iso_surface settings.  Not generally useful until I also add a step to
stitch the iso_surface pieces from adjacent elements together before
decimating.  Additionally many of the data structures need rework and
consolidation.


git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@4141 3c1deb5b-d424-0410-962d-aba41a686d42