Remove need for reverse spectrum colour mapping
Remove limitations on colour minimum and maximum to allow them to be swapped. Remove reverse mode from default spectrum and swap its minimum and maximum colours so it is still blue-to-red. Update API docs to discourage use of reverse mode. This is to permit clients to not use reverse, which makes the colour ranges work non-intuitively.