Wrap non-RC fields when set in graphics
Previously wrappers for non-Rectangular Cartesian (non-RC) coordinate and vector fields in graphics were silently created and destroyed as graphics objects were generated. To prevent notifications, this was done between calls to fieldmodule begin/end change, which disabled field caching so evaluations were less efficient. It also exposed the flaw in find xi caching which is still to be fixed.
After this change, whenever a coordinate or vector field is set in graphics, these are registered with the scene and an RC wrapper is created if appropriate. On the downside this means these wrapper fields are visible (with names such as <original_field_name>_cmiss_rc_wrapper) but unmanaged, however this solution is more efficient will make it easier to parallelise later.
Issue #50