Stop caching last location in find xi code
Field cache already handles caching of last value.
Also avoid searching last element in 'find nearest' mode since all elements need to be searched in that case.
There are still issues with the find xi cache which maintains a non-accessed pointer to the last element, but this will need to be addressed later.
Issue #50.